Added support for parsing playlists and albums
Playlists now parsed, although not all playlist
data can be viewed in a human readale format, it's
mostly just how it was read from the file.

Discovered this whole 'albums' thing and added support
for parsing those.

A number of new classes to support all of this, and
some changes to Utils to prevent croaking when I request
to read 0 bytes and then actually do ;)

NOTE: This is all still extremely beta and not the
fastest thing in the world, but you can, for instance,
write a fairly simple script to get all songs from the
iPod and copy them onto your disk.

I also doubt the dist.ini file for this project will
actually succeed in building anything. It certainly
won't pass any tests...

...which reminds me, I still haven't written any tests...
I'll get around to that soon enough.